Install dokku-slack plugin
This plugin required having environment variables exported in
/home/dokku/dokkurc. The commit gives each plugin an env key which can
be used to define key/val pairs that will be exported. Global env
variables can be set through an env key on the role, this can be used to
set DOKKU_TRACE.

This also introduces the first vault file for storing the slack_url
variable, which points at the slack incoming webhook url.